Implemented storage of Week Availability in the Instructor and Location databases.
git-svn-id: svn+ssh://scheduler.csc.calpoly.edu/repo@390 13b268f7-8392-4c9d-bc7d-1e80a86bc2f9

+++ implementation/model/java/edu/calpoly/csc/scheduler/model/schedule/DayAvail.java
@@ -1,9 +1,8 @@
 package edu.calpoly.csc.scheduler.model.schedule;
 
 
-import java.util.BitSet;
-import java.util.Vector;
 import java.io.Serializable;
+import java.util.BitSet;
 
 import edu.calpoly.csc.scheduler.model.db.Time;
 
@@ -15,7 +14,7 @@ import edu.calpoly.csc.scheduler.model.db.Time;
  */
 public class DayAvail extends GenAvail<Void> implements Serializable
 {
-   public static final int serialVersionUID = 42;
+   public static final long serialVersionUID = 42;
 
    public DayAvail ()/*==>*/
    {
